requirements.txt
langchain
langchain_groq
psycopg2-binary
python-dotenv
sqlalchemy
streamlit
langchain-community